make etasu component modular and add to patient portal
