+++ src/components/EtasuStatus/EtasuStatus.jsx
@@ -1,51 +1,44 @@
 import { EtasuStatusButton } from './EtasuStatusButton.jsx';
 import { EtasuStatusModal } from './EtasuStatusModal.jsx';
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useContext } from 'react';
 import { Card, Typography } from '@mui/material';
+import { SettingsContext } from '../../containers/ContextProvider/SettingsProvider.jsx';
+import axios from 'axios';
+import { EtasuStatusComponent } from './EtasuStatusComponent.jsx';
+import { getEtasu } from '../../util/util.js';
 
+// converts code into etasu for the component to render
+// simplifies usage for applications that only know the code, not the case they want to display
 export const EtasuStatus = props => {
-  const { etasuUrl, request, remsAdminResponse, getEtasuStatus, lastCheckedEtasuTime } =
+  const [globalState, _] = useContext(SettingsContext);
+
+  const { code } =
     props;
-  const [showEtasuStatus, setShowEtasuStatus] = useState(false);
+  const [remsAdminResponse, setRemsAdminResponse] = useState({});
+  useEffect(() => getEtasuStatus(), [code]);
+  const getEtasuStatus = () => {
+    const patientFirstName = globalState.patient?.name?.at(0)?.given?.at(0);
+    const patientLastName = globalState.patient?.name?.at(0)?.family;
+    const patientDOB = globalState.patient?.birthDate;
 
-  useEffect(() => getEtasuStatus(), [request.id, etasuUrl]);
-
-  const handleCloseEtasuStatus = () => {
-    setShowEtasuStatus(false);
-  };
-
-  const handleOpenEtasuStatus = () => {
-    setShowEtasuStatus(true);
+    console.log(
+      'get Etastu Status: ' +
+        patientFirstName +
+        ' ' +
+        patientLastName +
+        ' - ' +
+        patientDOB +
+        ' - ' +
+        code
+    );
+    const etasuUrl = `${globalState.remsAdminServer}/etasu/met/patient/${patientFirstName}/${patientLastName}/${patientDOB}/drugCode/${code}`;
+    getEtasu(etasuUrl, setRemsAdminResponse);
   };
 
   return (
-    <Card variant="outlined" sx={{ padding: 2 }}>
-      <Typography variant="h6" align="center" mb={2}>
-        {remsAdminResponse?.drugName}
-      </Typography>
-      <EtasuStatusButton
-        baseColor={getStatusColor(remsAdminResponse?.status)}
-        remsAdminResponse={remsAdminResponse}
-        handleOpenEtasuStatus={handleOpenEtasuStatus}
-        lastCheckedEtasuTime={lastCheckedEtasuTime}
-      />
-      <EtasuStatusModal
-        callback={getEtasuStatus}
-        remsAdminResponse={remsAdminResponse}
-        update={showEtasuStatus}
-        onClose={handleCloseEtasuStatus}
-      />
-    </Card>
+    <>
+      {remsAdminResponse.case_number ? <EtasuStatusComponent remsAdminResponseInit={remsAdminResponse} /> : ""}
+    </>
   );
 };
 
-export const getStatusColor = status => {
-  switch (status) {
-    case 'Approved':
-        return 'green';
-    case 'Pending':
-        return '#f0ad4e';
-    default:
-      return '#0c0c0c';
-  }
-};
